requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
pillow>=9.0.0
opencv-python>=4.7.0
python-dotenv>=1.0.0
psutil>=5.9.0
GPUtil>=1.4.0
dashscope>=1.14.0
numpy>=1.24.0
matplotlib>=3.6.0
tqdm>=4.64.0
requests>=2.28.0
urllib3>=1.26.0
certifi>=2022.12.7

# 좌표 매핑을 위한 선택적 OCR 라이브러리 (더 정확한 결과를 원한다면 설치)
# easyocr>=1.6.0
